Split tests into deterministic and Hypothesis-based
