+++ tests/test_deterministic/test_constants.py
@@ -19,9 +19,6 @@
 # isort: STDLIB
 import unittest
 
-# isort: THIRDPARTY
-from hypothesis import given, strategies
-
 # isort: LOCAL
 from justbytes._constants import UNITS, B, BinaryUnits, DecimalUnits, RoundingMethods
 from justbytes._errors import RangeValueError
@@ -44,19 +41,6 @@ class ConstantsTestCase(unittest.TestCase):
         self.assertTrue(set(DecimalUnits.UNITS()).issubset(set(UNITS())))
         self.assertTrue(set(BinaryUnits.UNITS()).issubset(set(UNITS())))
         self.assertTrue(B in UNITS())
-
-    @given(
-        strategies.integers(min_value=0, max_value=BinaryUnits.max_exponent()),
-        strategies.integers(min_value=0, max_value=DecimalUnits.max_exponent()),
-    )
-    def testExpMethod(self, bexp, dexp):
-        """ Test extracting unit for a given exponent. """
-        self.assertEqual(
-            BinaryUnits.unit_for_exp(bexp).factor, BinaryUnits.FACTOR ** bexp
-        )
-        self.assertEqual(
-            DecimalUnits.unit_for_exp(dexp).factor, DecimalUnits.FACTOR ** dexp
-        )
 
     def testExpExceptions(self):
         """
