+++ tests/test_deterministic/test_errors.py
