+++ tests/test_deterministic/test_size/test_size.py
